% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_journeys.R
\name{get_journeys}
\alias{get_journeys}
\title{get_jouneys}
\usage{
get_journeys(start_date = Sys.Date() - 1, end_date = Sys.Date())
}
\value{
Number of records updated
}
\description{
Retrieve journey records between two dates.
}
