% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_graphql_request.R
\name{graphql_request}
\alias{graphql_request}
\title{graphql_request}
\usage{
graphql_request(query, variables)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
