% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_journeys.R
\name{get_journeys}
\alias{get_journeys}
\title{get_data}
\usage{
get_journeys(conn, url, project_id, start_id)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
